Create 'tools' and move playit
- Created 'tools' for helpers for Minecraft, that don't relate to the game itself. For instance, Java, playit, and Telepath (even though it's integrated)
