+++ source/core/server/manager.py
@@ -21,8 +21,9 @@ import re
 
 from source.core.server.acl import AclManager, get_uuid, check_online
 from source.core.server.backup import BackupManager
-from source.core.server import backup, playit
 from source.core import constants, telepath
+from source.core.server import backup
+from source.core.tools import playit
 from source.core.constants import (
 
     # Directories
