+++ source/ui/desktop/views/server/manager/settings.py
@@ -1,6 +1,6 @@
 from ui.desktop.views.server.manager.editor import open_config_file
 from source.ui.desktop.views.server.manager.components import *
-from source.core.server import playit
+from source.core.tools import playit
 
 
 # ---------------------------------------------- Server Settings Screen ------------------------------------------------
